// src/models/pagination.model.ts
export class PaginationModel {
    id: number;
    page: any = 0;
    size = 12;
    keyword: string;
    user_type: string;
    farmer_id: string;
    category_id: string;
    order_id: string;
    kitchen_id: number;
    type: number;
    address_id: number;
    status: number;
    endDate: string;
    startDate: string;
    max_percentage: string;
    image: string;
}